pyproject.toml
[project]
name = "bookingagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "langchain-core>=1.2.14",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.3.0",
    "python-dotenv>=1.2.1",
]
